Handling the BottomBar color change with the new tabs properly.
