+++ bottom-bar/src/main/java/com/roughike/bottombar/MiscUtils.java
@@ -101,85 +101,6 @@ class MiscUtils {
     }
 
     /**
-     * Animate a background activeIconColor change. Uses Circular Reveal if supported,
-     * otherwise crossfades the background activeIconColor in.
-     *
-     * @param clickedView    the view that was clicked for calculating the start position
-     *                       for the Circular Reveal.
-     * @param backgroundView the currently showing background activeIconColor.
-     * @param bgOverlay      the overlay view for the new background activeIconColor that will be
-     *                       animated in.
-     * @param newColor       the new activeIconColor.
-     */
-    @TargetApi(Build.VERSION_CODES.LOLLIPOP)
-    protected static void animateBGColorChange(View clickedView, final View backgroundView,
-                                               final View bgOverlay, final int newColor) {
-        int centerX = (int) (ViewCompat.getX(clickedView) + (clickedView.getMeasuredWidth() / 2));
-        int centerY = clickedView.getMeasuredHeight() / 2;
-        int finalRadius = backgroundView.getWidth();
-
-        backgroundView.clearAnimation();
-        bgOverlay.clearAnimation();
-
-        Object animator;
-
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
-            if (!bgOverlay.isAttachedToWindow()) {
-                return;
-            }
-
-            animator = ViewAnimationUtils
-                    .createCircularReveal(bgOverlay, centerX, centerY, 0, finalRadius);
-        } else {
-            ViewCompat.setAlpha(bgOverlay, 0);
-            animator = ViewCompat.animate(bgOverlay).alpha(1);
-        }
-
-        if (animator instanceof ViewPropertyAnimatorCompat) {
-            ((ViewPropertyAnimatorCompat) animator).setListener(new ViewPropertyAnimatorListenerAdapter() {
-                @Override
-                public void onAnimationEnd(View view) {
-                    onCancel();
-                }
-
-                @Override
-                public void onAnimationCancel(View view) {
-                    onCancel();
-                }
-
-                private void onCancel() {
-                    backgroundView.setBackgroundColor(newColor);
-                    bgOverlay.setVisibility(View.INVISIBLE);
-                    ViewCompat.setAlpha(bgOverlay, 1);
-                }
-            }).start();
-        } else if (animator != null) {
-            ((Animator) animator).addListener(new AnimatorListenerAdapter() {
-                @Override
-                public void onAnimationEnd(Animator animation) {
-                    onCancel();
-                }
-
-                @Override
-                public void onAnimationCancel(Animator animation) {
-                    onCancel();
-                }
-
-                private void onCancel() {
-                    backgroundView.setBackgroundColor(newColor);
-                    bgOverlay.setVisibility(View.INVISIBLE);
-                    ViewCompat.setAlpha(bgOverlay, 1);
-                }
-            });
-
-            ((Animator) animator).start();
-        }
-
-        bgOverlay.setBackgroundColor(newColor);
-        bgOverlay.setVisibility(View.VISIBLE);
-    }
-
-    /**
      * A convenience method for setting text appearance.
      *
      * @param textView a TextView which textAppearance to modify.
